refactor(hooks): useRoastTimerを4つのカスタムフックに分割
- useTimerState: 状態管理とRef管理
- useTimerControls: タイマー操作と完了処理
- useTimerNotifications: 音声・通知管理
- useTimerPersistence: Firestore/LocalStorage同期

996行のモノリシックなフックを責務ごとに分割し、
テスト容易性と可読性を大幅に改善。
外部インターフェースは完全に互換性を維持。

Closes #36
